Balanced Team Problem
You are a coach at your local university. There are n students under your supervision, the programming skill of the i-th student is a[i].
You have to create a team for a new programming competition. As you know, the more students some team has, the more probable its victory is.
So you have to create a team with the maximum number of students. But you also know that a team should be balanced. It means that the programming skill of each pair of students in a created team should differ by no more than 5.
Your Task is to report the maximum possible number of students in a Balanced Team.

Input
The first line of the input contains one integer n(1&lt;=n&lt;=10^5) - the number of students.
The second line of the input contains n integers a1,a2,...,an n(1&lt;=a[i]&lt;=10^9), where ai is a programming skill of the i-th student.
Output
Print one integer - the maximum possible number of students in a balanced team.

<strong>Examples:</strong>
Input :
6 
1 10 17 12 15 2
Output :
3
Input :
3
1 100 100000